Internship/requirements.txt
numpy~=2.1.3
torchvision~=0.20.1
pandas~=2.2.3
pillow~=11.0.0
scikit-learn~=1.5.2
typing_extensions~=4.12.2
zlib~=1.2.13
pip~=24.2
wheel~=0.44.0
pysocks~=1.7.1
openssl~=3.4.0
tornado~=6.4.2
jinja2~=3.1.4
setuptools~=72.1.0
pytz~=2024.1
charset-normalizer~=3.3.2
sympy~=1.13.3
mpmath~=1.3.0
gmpy2~=2.1.2
python~=3.11.11
matplotlib~=3.9.2
tqdm~=4.67.1
scipy~=1.14.1
packaging~=24.2
colorama~=0.4.6
filelock~=3.13.1
urllib3~=2.2.3
requests~=2.32.3
markupsafe~=2.1.3
networkx~=3.2.1
yaml~=0.2.5
pyyaml~=6.0.2
zipp~=3.21.0
torchaudio~=2.5.1
optuna~=4.1.0
protobuf~=5.29.3
flask~=3.1.0
werkzeug~=3.1.3
flask-jwt-extended~=4.7.0